refactor: replace readFile function to file module
