+++ src/modules/file.js
@@ -1,5 +1,22 @@
 import fsPromises from 'node:fs/promises'
-import path from 'node:path'
+import fs from 'node:fs'
+import path from 'path'
+
+const readFile = async (pathToFile) => {
+  try {
+    await new Promise((resolve, reject) => {
+      const readableStream = fs.createReadStream(
+        path.join(process.cwd(), pathToFile),
+        'utf-8'
+      )
+      readableStream.on('data', (chunk) => console.log(chunk))
+      readableStream.on('end', () => resolve())
+      readableStream.on('error', () => reject())
+    })
+  } catch (error) {
+    throw new Error()
+  }
+}
 
 const createFile = async (fileName) => {
   let fileHandle
@@ -14,4 +31,4 @@ const createFile = async (fileName) => {
   }
 }
 
-export { createFile }
+export { createFile, readFile }
